feat: implement most of the lexer
so as i suspected, this was super duper easy lol. i just kinda set myself up for
the layup and then hit that shit. hell yeah man. this is super cool. because now
its like a super simple implementation (and probably super fast too lol idk
tho). well it should be pretty simple to now go and implement the rest of the
features. cool.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>