chore: finish array list refactor
this makes it use an explicit length value and memcpy in extend, so there should
be overall a shitload less branches overall and hopefully that also means
quicker compile times.

could also refactor how the map works and probably should. the regex module also
needs a HUGE refactor but i am honestly not ready for that.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>