feat: even more halfway through the refactor
okay this is working perfectly, in fact, at the root level. all that's really
left is to make the things that should expand actually expand.

but better yet, this implementation makes way more sense AND actually works
better and supports the things i want. uhmmm there are definitely still cases
where it doesnt really do what it should, but its doing what i need it to and
thats the name of the game babye

Signed-off-by: Matt Provost <mattprovost6@gmail.com>