feat: add basic benchmark stats
so this is super rudimentary and it would honestly be pretty cool to actually
build out better benchmarking stuff to kinda run the code like hyperfine does
and generate stddev and all that kinda stuff. would be pretty cool to see the
better average stats and all that.

anyway, also did some minor readme updates and stuff.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>